Rename core modules to PascalCase
