+++ TradingBot/core/Branch.py
