+++ TradingBot/core/datamanager.py
@@ -7,8 +7,12 @@ import pandas as pd
 
 from TradingBot.brokers.ExchangeBybit import Exchange_Bybit as ExchangeAPI
 from TradingBot.core.configmanager import config_manager
-from TradingBot.core.utilities import (ensure_directory, format_timestamp, retry,
-                                      write_json)
+from TradingBot.core.Utilities import (
+    ensure_directory,
+    format_timestamp,
+    retry,
+    write_json,
+)
 
 logger = logging.getLogger(__name__)
